counts modifications for ui

+++ src/resolvers/Count/Count.test.ts
@@ -19,7 +19,7 @@ afterAll(async () => {
 });
 
 const getAllCountsQuery = `
-    query Query($exerciseId: Float) {
+    query Query($exerciseId: Int) {
         counts(exerciseId: $exerciseId) {
             id
             date
